+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 * Icon Button: Nieuw component ([#2315](https://github.com/dso-toolkit/dso-toolkit/issues/2315))
+* Tijdreis Banner: Nieuw component ([#3246](https://github.com/dso-toolkit/dso-toolkit/issues/2346))
 
 ## 🧚‍♀️ Release 79.0.1 - 2025-09-22